django-backend/requirements.txt
# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery for async tasks
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Redis client
redis>=5.0,<6.0

# Cryptography
cryptography>=41.0,<43.0
pycryptodome>=3.18,<4.0

# Stellar SDK
stellar-sdk>=11.0,<12.0

# ZK Proofs (placeholder - will need rust bindings)
# bulletproofs>=0.1.0

# Environment variables
python-dotenv>=1.0,<2.0

# Validation
pydantic>=2.0,<3.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Utilities
python-dateutil>=2.8,<3.0